requirements.txt
fastapi[all]
SQLAlchemy==1.4.41
requests==2.28.1
python-dotenv==0.21.0
alembic==1.8.1
asyncpg==0.26.0
lxml==4.9.1
